enhancement(HeroBackground): Improve crop & resize behavior
