+++ components/collective-page/hero/Hero.js
@@ -24,6 +24,7 @@ import { Flex } from '../../Grid';
 import I18nCollectiveTags from '../../I18nCollectiveTags';
 import Link from '../../Link';
 import LinkCollective from '../../LinkCollective';
+import LoadingPlaceholder from '../../LoadingPlaceholder';
 import MessageBox from '../../MessageBox';
 import StyledButton from '../../StyledButton';
 import StyledLink from '../../StyledLink';
@@ -31,16 +32,25 @@ import StyledRoundButton from '../../StyledRoundButton';
 import StyledTag from '../../StyledTag';
 import { H1, Span } from '../../Text';
 import UserCompany from '../../UserCompany';
-// Local imports
 import ContainerSectionContent from '../ContainerSectionContent';
 
 import CollectiveColorPicker from './CollectiveColorPicker';
 import HeroAvatar from './HeroAvatar';
-import HeroBackground from './HeroBackground';
+import HeroBackground, { BASE_HERO_HEIGHT, StyledHeroBackground } from './HeroBackground';
 import HeroTotalCollectiveContributionsWithData from './HeroTotalCollectiveContributionsWithData';
 
 // Dynamic imports
 const HeroEventDetails = dynamic(() => import('./HeroEventDetails'));
+
+const HeroBackgroundEdit = dynamic(() => import('./HeroBackgroundEdit'), {
+  loading() {
+    return (
+      <StyledHeroBackground>
+        <LoadingPlaceholder height={BASE_HERO_HEIGHT} />
+      </StyledHeroBackground>
+    );
+  },
+});
 
 const Translations = defineMessages({
   website: {
@@ -97,7 +107,12 @@ const Hero = ({ collective, host, isAdmin, onPrimaryColorChange, callsToAction, 
         </MessageBox>
       )}
       <Container position="relative" minHeight={325} zIndex={1000} data-cy="collective-hero">
-        <HeroBackground collective={collective} isEditing={isEditingCover} onEditCancel={() => editCover(false)} />
+        {isEditing ? (
+          <HeroBackgroundEdit collective={collective} onEditCancel={() => editCover(false)} />
+        ) : (
+          <HeroBackground collective={collective} />
+        )}
+
         {isAdmin && !isEditing && (
           // We don't have any mobile view for this one yet
           <Container
